Migrated f1.py
